airflow/requirements4.txt
# aiohttp==3.9.3
# aiosignal==1.3.1
# alembic==1.13.1
annotated-types==0.6.0
# anyio==4.3.0
# apache-airflow==2.8.4
# apache-airflow-providers-common-io==1.3.0
# apache-airflow-providers-common-sql==1.11.1
# apache-airflow-providers-ftp==3.7.0
# apache-airflow-providers-http==4.10.0
# apache-airflow-providers-imap==3.5.0
# apache-airflow-providers-smtp==1.6.1
# apache-airflow-providers-sqlite==3.7.1
# apispec==6.6.0
# argcomplete==3.2.3
asgiref==3.8.1
asttokens==2.4.1
# async-timeout==4.0.3
# attrs==23.2.0
# Babel==2.14.0
blinker==1.4
# cachelib==0.9.0
# certifi==2024.2.2
# cffi==1.16.0
chardet==3.0.4
# charset-normalizer==3.3.2
# click==8.1.7
# clickclick==20.10.2
# colorama==0.4.6
# colorlog==4.8.0
comm==0.2.1
# ConfigUpdater==3.2
# connexion==2.14.2
# cron-descriptor==1.4.3
# croniter==2.0.3
cryptography==3.4.8
Cython==3.0.8
decorator==5.1.1
Deprecated==1.2.14
# dill==0.3.8
distlib==0.3.4
distro==1.7.0
# dnspython==2.6.1
# docutils==0.20.1
# email-validator==1.3.1
# exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
# Flask==2.2.5
# Flask-AppBuilder==4.3.11
# Flask-Babel==2.0.0
# Flask-Caching==2.1.0
# Flask-JWT-Extended==4.6.0
# Flask-Limiter==3.5.1
# Flask-Login==0.6.3
# Flask-Session==0.5.0
# Flask-SQLAlchemy==2.5.1
# Flask-WTF==1.2.1
# frozenlist==1.4.1
fsspec==2024.2.0
gluonnlp==0.10.0
# google-re2==1.1
# googleapis-common-protos==1.63.0
graphviz==0.8.4
# greenlet==3.0.3
# grpcio==1.62.1
# gunicorn==21.2.0
# h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.2
# idna==3.6
importlib-metadata==4.6.4
importlib_resources==6.4.0
# inflection==0.5.1
ipython==8.18.1
ipywidgets==8.1.2
# itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
# Jinja2==3.1.3
joblib==1.3.2
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
keyring==23.5.0
kobert-tokenizer==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
# limits==3.10.1
# linkify-it-py==2.0.3
# lockfile==0.12.2
# Mako==1.3.2
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# marshmallow==3.21.1
# marshmallow-oneofschema==3.1.1
# marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
# mdit-py-plugins==0.4.0
# mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
# multidict==6.0.5
mxnet==1.7.0.post2
netifaces==0.11.0
networkx==3.2.1
numpy==1.23.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.0
# opentelemetry-api==1.23.0
# opentelemetry-exporter-otlp==1.23.0
# opentelemetry-exporter-otlp-proto-common==1.23.0
# opentelemetry-exporter-otlp-proto-grpc==1.23.0
# opentelemetry-exporter-otlp-proto-http==1.23.0
# opentelemetry-proto==1.23.0
# opentelemetry-sdk==1.23.0
# opentelemetry-semantic-conventions==0.44b0
# ordered-set==4.1.0
packaging==23.2
pandas==2.1.0
parso==0.8.3
# pathspec==0.12.1
# pendulum==3.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==2.5.1
# pluggy==1.4.0
# prison==0.2.1
prompt-toolkit==3.0.43
# protobuf==4.25.3
# psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.26.0
# pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
# Pygments==2.17.2
PyJWT==2.3.0
PyMySQL==1.1.0
pyparsing==2.4.7
# python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
# python-nvd3==0.15.0
# python-slugify==8.0.4
# pytz==2024.1
# PyYAML==6.0.1
# referencing==0.34.0
regex==2023.12.25
# requests==2.31.0
# requests-toolbelt==1.0.0
# rfc3339-validator==0.1.4
rich==13.7.1
# rich-argparse==1.4.0
# rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
sentencepiece==0.2.0
# setproctitle==1.3.3
six==1.16.0
# sniffio==1.3.1
SQLAlchemy==1.4.49
# SQLAlchemy-JSONField==1.0.2
# SQLAlchemy-Utils==0.41.2
# sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
# tabulate==0.9.0
# tenacity==8.2.3
# termcolor==2.4.0
# text-unidecode==1.3
threadpoolctl==3.3.0
time-machine==2.14.1
tokenizers==0.15.2
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
triton==2.2.0
# typing_extensions==4.10.0
# tzdata==2024.1
# uc-micro-py==1.0.3
# unicodecsv==0.14.1
# universal-pathlib==0.1.4
urllib3==2.2.1
uvicorn==0.28.0
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
# Werkzeug==2.2.3
widgetsnbextension==4.0.10
# wrapt==1.16.0
# WTForms==3.1.2
# yarl==1.9.4
zipp==1.0.0
